requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
requests==2.31.0
Pillow==10.0.1
pygame==2.5.2
schedule==1.2.0
python-dotenv==1.0.0
